+++ CHANGELOG.md
@@ -7,13 +7,14 @@
 * Various optimizations and improvements to mathematical functions.
 * Generalize the prime sieve implementation and add support for more efficient primality check and factorization.
 * More efficient implementation of `repeat` for a single element.
-* Deprecate `Iterable.cycle` in favor of `Iterable.repeat`.
 * Make `Iterable.groupBy` and `Iterable.indexed` return instances of `MapEntry` for easier interoperability with `Map`.
 * Fix various edge cases with `Range` iterables.
-* Deprecated `naturalComparator` in favor of `naturalCompare` and `naturalComparable<T>`.
+* Fix `Iterable.combinations` for empty iterables.
 * Add `List` extensions `takeTo`, `skipTo`, `takeLast`, `takeLastTo`, `skipLast`, and `skipLastTo`.
+* Add `Iterable` extension `count` and `powerSet`.
+* Deprecate `Iterable.cycle` in favor of `Iterable.repeat`.
+* Deprecated `naturalComparator` in favor of `naturalCompare` and `naturalComparable<T>`.
 * Deprecate the iterable package, and integrate the extensions into the collection package.
-* Add `Iterable` extension `count`.
 
 ## 3.8.0
 * Add `String` extensions `take`, `takeTo`, `skip`, `skipTo`, `takeLast`, `takeLastTo`, `skipLast`, and `skipLastTo`.